BIT/requirements.txt
python~=3.9.13
pysnowflake~=0.1.3
snowflake-snowpark-python~=1.8.0
pandas~=2.1.1